requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
aiosqlite==0.20.0
alembic==1.13.3
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
nba_api==1.11.4
pandas==2.2.3
google-generativeai==0.8.3
python-dotenv==1.0.1
email-validator==2.2.0
curl_cffi==0.7.1
